backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio
redis
redis.asyncio
passlib[bcrypt]
python-jose[cryptography]
